% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-estimate_coverage.R
\name{calculate_cf}
\alias{calculate_cf}
\title{Estimate case finding effectiveness}
\usage{
calculate_cf(cin, cout)
}
\arguments{
\item{cin}{Cases in CMAM programme}

\item{cout}{Cases not in CMAM programme}
}
\value{
Value of case finding effectiveness
}
\description{
Estimate case finding effectiveness
}
\examples{
calculate_cf(cin = 5, cout = 20)

}
