% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calculate_cured}
\alias{calculate_cured}
\title{Calculate CMAM performance indicators - cure rate}
\usage{
calculate_cured(cured, exit)
}
\arguments{
\item{cured}{Numeric value for total number of cured cases}

\item{exit}{Numeric value for total number of programme exits}
}
\value{
A proportion of cured cases of the total programme exits
}
\description{
Calculate CMAM performance indicators - cure rate
}
\examples{
calculate_cured(cured = 10, exit = 50)

}
\author{
Ernest Guevarra
}
