% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calculate_dead}
\alias{calculate_dead}
\title{Calculate CMAM performance indicators - death rate}
\usage{
calculate_dead(dead, exit)
}
\arguments{
\item{dead}{Numeric value for total number of cases who died}

\item{exit}{Numeric value for total number of programme exits}
}
\value{
A proportion of dead cases of the total programme exits
}
\description{
Calculate CMAM performance indicators - death rate
}
\examples{
calculate_dead(dead = 10, exit = 50)

}
\author{
Ernest Guevarra
}
