% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calculate_default}
\alias{calculate_default}
\title{Calculate CMAM performance indicators - default rate}
\usage{
calculate_default(defaulter, exit)
}
\arguments{
\item{defaulter}{Numeric value for total number of cases who defaulted}

\item{exit}{Numeric value for total number of programme exits}
}
\value{
A proportion of defaulter cases of the total programme exits
}
\description{
Calculate CMAM performance indicators - default rate
}
\examples{
calculate_default(defaulter = 10, exit = 50)

}
\author{
Ernest Guevarra
}
