% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_of_stay.R
\name{calculate_los_df}
\alias{calculate_los_df}
\title{Calculate length of stay for a cohort of CMAM discharges}
\usage{
calculate_los_df(df, admission_date, discharge_date, add = TRUE)
}
\arguments{
\item{df}{A data.frame containing information on a cohort of CMAM
discharges specifically dates of admission and dates of discharge.}

\item{admission_date}{A character value for name of variable in \code{df}
containing dates of admission in \code{YYYY-MM-DD} format.}

\item{discharge_date}{A character value for name of variable in \code{df}
containing dates of discharge in \code{YYYY-MM-DD} format.}

\item{add}{Logical. Should results be added as a new column in \code{df}? Default
is TRUE.}
}
\value{
If \code{add} is TRUE, a tibble created from column binding \code{df} with
the vector of numerical values of calculated length-of-stay. If \code{add} is
FALSE, a vectdor of numerical values of calculated length-of-stay.
}
\description{
Calculate length of stay for a cohort of CMAM discharges
}
\examples{
x <- data.frame(admission = c("2010-03-15", "2010-02-15"),
                discharge = c("2010-06-14", "2010-04-19"))

calculate_los_df(df = x,
                 admission_date = "admission",
                 discharge_date = "discharge")

}
\author{
Ernest Guevarra
}
