% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muac_at_admission.R
\name{calculate_muac_median}
\alias{calculate_muac_median}
\title{Calculate median MUAC-at_admissions}
\usage{
calculate_muac_median(muac, index = NULL, na_values = NULL)
}
\arguments{
\item{muac}{A numeric vector of mid-upper arm circumference measurements
either in centimetres or millimetres.}

\item{index}{A list of one or more factors each of the same length as \code{muac}
which defines the grouping to be used on \code{muac}. Default is NULL (no
grouping applied).}

\item{na_values}{A vector of values in \code{muac} that are to be considered as
NA values. Default to NULL which will use NA.}
}
\value{
A numeric value (if \code{index = NULL}) or named vector of values for
median mid-upper arm circumference with names from each level of \code{index}.
}
\description{
Calculate median MUAC-at_admissions
}
\examples{
calculate_muac_median(otp_beneficiaries$muac)
calculate_muac_median(otp_beneficiaries$muac, otp_beneficiaries$locality)

}
