% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calculate_no_response}
\alias{calculate_no_response}
\title{Calculate CMAM performance indicators - non-response rate}
\usage{
calculate_no_response(nr, exit)
}
\arguments{
\item{nr}{Numeric value for total number of cases who did not respond to
treatment}

\item{exit}{Numeric value for total number of programme exits}
}
\value{
A proportion of non-responders of the total programme exits
}
\description{
Calculate CMAM performance indicators - non-response rate
}
\examples{
calculate_no_response(nr = 10, exit = 50)

}
\author{
Ernest Guevarra
}
