% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calculate_performance}
\alias{calculate_performance}
\title{Calculate CMAM performance indicators}
\usage{
calculate_performance(.data, vars = NULL, add = TRUE)
}
\arguments{
\item{.data}{A data.frame containing programme monitoring data on number of
cured, deaths, defaulters and non-response. The required data.frame
holds rows of data corresponding to total counts for each performance
indicator (i.e., \strong{cured}, \strong{dead}, \strong{defaulter} and \strong{non-responder})
rather than rows of individual cases.}

\item{vars}{A vector of variable names specifying \strong{cured}, \strong{dead},
\strong{defaulter} and \strong{non-responder} (in this specific order) values in
\code{.data}. If NULL (default), typical names used for these variables will
be searched and used accordingly. If search doesn't yield matching variable
names, the first 4 columns of the data.frame will be used.}

\item{add}{Logical. Should result be added to \code{.data}. Default is TRUE.}
}
\value{
A tibble of performance indicator results
}
\description{
Calculate CMAM performance indicators
}
\examples{
calculate_performance(.data = monitoring)

}
\author{
Ernest Guevarra
}
