% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-estimate_coverage.R
\name{calculate_rout}
\alias{calculate_rout}
\title{Estimate cases not in CMAM programme}
\usage{
calculate_rout(cin, cout, rin, k = 3)
}
\arguments{
\item{cin}{Cases in CMAM programme}

\item{cout}{Cases not in CMAM programme}

\item{rin}{Recovering cases in programme}

\item{k}{Correction factor. Ratio of the mean length of an untreated episode
to the mean length of a CMAM treatment episode}
}
\value{
Value of number of cases not in CMAM progamme
}
\description{
Estimate cases not in CMAM programme
}
\examples{
calculate_rout(cin = 5, cout = 25, rin = 5, k = 3)

}
