% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_var_names}
\alias{find_var_names}
\title{Find possible variable names from a data.frame given a set of search names}
\usage{
find_var_names(df, vars, all = FALSE)
}
\arguments{
\item{df}{A data.frame to search variable names from}

\item{vars}{A vector of terms to search for}

\item{all}{Logical. Should all results of search be returned? If FALSE
(default), only first value is returned.}
}
\value{
A character vector of variable name/s in \code{df}
}
\description{
Find possible variable names from a data.frame given a set of search names
}
\examples{
find_var_names(df = otp_beneficiaries, vars = "MUAC")

}
\author{
Ernest Guevarra
}
