% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monitoring}
\alias{monitoring}
\title{Routine CMAM monitoring data from Sudan}
\format{
A tibble with 8234 rows and 16 columns\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{State} \tab Name of state \cr
   \emph{Locality} \tab Name of locality \cr
   \emph{Beginning of Month} \tab Cases in programme at beginning of month \cr
   \emph{New Admissions} \tab New cases admitted within the month \cr
   \emph{Male} \tab New male cases admitted within the month \cr
   \emph{Female} \tab New female cases admitted within the month \cr
   \emph{Cured} \tab Number of cured cases within the month \cr
   \emph{Death} \tab Number of cases who died within the month \cr
   \emph{Default} \tab Number of cases who defaulted within the month \cr
   \emph{Non-Responder} \tab Number of non-responder cases within the month \cr
   \emph{Total Discharge} \tab Total number of discharges within the month \cr
   \emph{RUTF Consumed} \tab Number of RUTF consumed \cr
   \emph{Screening} \tab Screening \cr
   \emph{Sites} \tab Sites \cr
   \emph{Month} \tab Month \cr
   \emph{Year} \tab Year \cr
}
}
\source{
Federal Ministry of Health of Sudan
}
\usage{
monitoring
}
\description{
Routine CMAM monitoring data from Sudan
}
\examples{
monitoring

}
\keyword{datasets}
