% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{muac_admission_tidy}
\alias{muac_admission_tidy}
\title{MUAC at admission in tidy format}
\format{
A tibble with 506 rows and 3 columns\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{muac} \tab Mid-upper arm circumference in centimetres \cr
   \emph{district} \tab Name of district \cr
   \emph{count} \tab Number of cases with specific MUAC \cr
}
}
\source{
A SQUEAC survey in Lokori, Kenya
}
\usage{
muac_admission_tidy
}
\description{
MUAC at admission in tidy format
}
\examples{
muac_admission

}
\keyword{datasets}
