% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otp_beneficiaries}
\alias{otp_beneficiaries}
\title{Outpatient Therapeutic Care Programme (OTP) beneficiaries data}
\format{
A tibble with 151 rows and 13 columns:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{index} \tab Unique identifier \cr
   \emph{state} \tab Name of state \cr
   \emph{locality} \tab Name of locality \cr
   \emph{health_facility} \tab Name of health facility \cr
   \emph{age} \tab Age of child \cr
   \emph{muac} \tab Mid-upper arm circumference (cms) at admission \cr
   \emph{wt} \tab Weight (kgs) at admission \cr
   \emph{ht} \tab Height (cms) at admission \cr
   \emph{admDate} \tab Date of admission \cr
   \emph{disDate} \tab Date of discharge \cr
   \emph{diswt} \tab Weight (kgs) at discharge \cr
   \emph{attended} \tab Number of OTP sessions attended \cr
   \emph{exitType} \tab Type of exit (cured, dead, default or non-responder) \cr
}
}
\source{
Data collected from beneficiary cards from Kassala State, Sudan
}
\usage{
otp_beneficiaries
}
\description{
Outpatient Therapeutic Care Programme (OTP) beneficiaries data
}
\examples{
otp_beneficiaries

}
\keyword{datasets}
