% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seasonal_calendar}
\alias{seasonal_calendar}
\title{Seasonal calendar data for Sudan}
\format{
A tibble with 28 rows and 4 columns\tabular{ll}{
   \strong{Variables} \tab \strong{Description} \cr
   \emph{event} \tab Name of seasonal calendar event or activity \cr
   \emph{start} \tab Starting date of event/activity \cr
   \emph{end} \tab Starting date of event/activity \cr
   \emph{group} \tab Classification/group of activity or event \cr
}
}
\source{
https://fews.net/east-africa/sudan/seasonal-calendar/december-2013
}
\usage{
seasonal_calendar
}
\description{
Seasonal calendar data for Sudan
}
\examples{
seasonal_calendar

}
\keyword{datasets}
