% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{smooth_m3a3}
\alias{smooth_m3a3}
\title{Apply median of 3 and average of 3 smoothing on a time series}
\usage{
smooth_m3a3(x)
}
\arguments{
\item{x}{A vector of numerical information to be smoothed}
}
\value{
A vector of smoothed data
}
\description{
Apply median of 3 and average of 3 smoothing on a time series
}
\examples{
x <- aggregate(cbind(`New Admissions`, Default) ~ Month + Year,
               data = monitoring, FUN = sum)
smooth_m3a3(x = x$Default)

}
\author{
Ernest Guevarra
}
