% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_to_travel}
\alias{time_to_travel}
\title{Time-to-travel to health facilities for beneficiaries and volunteers}
\format{
A tibble with 117 rows and 9 columns:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{State} \tab Name of state \cr
   \emph{Locality} \tab Name of locality \cr
   \emph{Health Facility} \tab Name of health facility \cr
   \emph{Category} \tab Category of beneficiary or volunteer \cr
   \emph{30 or less} \tab Travel time of 30 minutes or less \cr
   \emph{31 to 60} \tab Travel time of 31 minutes to 60 minutes \cr
   \emph{61 to 90} \tab Travel time of 61 minutes to 90 minutes \cr
   \emph{91 to 120} \tab Travel time of 91 minutes to 120 minutes \cr
   \emph{more than 120} \tab Travel time of more than 120 minutes \cr
}
}
\source{
Data collected from beneficiary cards from Kassala State, Sudan
}
\usage{
time_to_travel
}
\description{
Time-to-travel to health facilities for beneficiaries and volunteers
}
\examples{
time_to_travel

}
\keyword{datasets}
