% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coverage.R
\name{calculate_cf}
\alias{calculate_cf}
\title{Estimate case finding effectiveness}
\usage{
calculate_cf(cin, cout)
}
\arguments{
\item{cin}{Cases in CMAM programme}

\item{cout}{Cases not in CMAM programme}
}
\value{
Value of case finding effectiveness
}
\description{
Estimate case finding effectiveness
}
\examples{
calculate_cf(cin = 5, cout = 20)

}
\references{
Safari Balegamire, Katja Siling, Jose Luis Alvarez Moran,
  Ernest Guevarra, Sophie Woodhead, Alison Norris, Lionella Fieschi,
  Paul Binns, and Mark Myatt (2015). A single coverage estimator for use in
  SQUEAC, SLEAC, and other CMAM coverage assessments. Field Exchange 49,
  March 2015. p81. <www.ennonline.net/fex/49/singlecoverage>
}
\author{
Ernest Guevarra based on technical notes and equations by Mark Myatt
}
