% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_of_stay.R
\name{calculate_los}
\alias{calculate_los}
\title{Calculate CMAM length of stay}
\usage{
calculate_los(admission, discharge)
}
\arguments{
\item{admission}{Date of admission in \code{YYYY-MM-DD} format. If child is a
kwashiorkor case, date of lowest weight (when oedema has subsided).}

\item{discharge}{Date of discharge in \code{YYYY-MM-DD} format.}
}
\value{
Numeric value for length-of-stay in days.
}
\description{
Calculate CMAM length of stay
}
\examples{
calculate_los(admission = "2010-03-15",
              discharge = "2010-06-14")

}
\author{
Ernest Guevarra
}
