% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_of_stay.R
\name{calculate_median_los}
\alias{calculate_median_los}
\title{Calculate median length of stay for a cohort of CMAM discharges}
\usage{
calculate_median_los(admission_date, discharge_date, group = NULL)
}
\arguments{
\item{admission_date}{Date of admission in \code{YYYY-MM-DD} format. If child is a
kwashiorkor case, date of lowest weight (when oedema has subsided). Can be
a single date value or a vector of date values.}

\item{discharge_date}{Date of discharge in \code{YYYY-MM-DD} format. Can be a
single date value or a vector of date values.}

\item{group}{A character value/s with the same length as \code{admission_date} and
\code{discharge_data} to use as grouping variable within which median
length-of-stay is to be calculated. Default is NULL for no grouping.}
}
\value{
A numeric value for median length-of-stay in days. If \code{group} is not
NULL, a vector of numeric values for median length-of-stay in days with
length equal to the number of groups.
}
\description{
Calculate median length of stay for a cohort of CMAM discharges
}
\examples{
calculate_median_los(
  otp_beneficiaries$admDate,
  otp_beneficiaries$disDate,
  group = otp_beneficiaries$locality
)

}
\author{
Ernest Guevarra
}
