% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-estimate_coverage.R
\name{calculate_tc}
\alias{calculate_tc}
\title{Estimate treatment coverage}
\usage{
calculate_tc(cin, cout, rin, k = 3)
}
\arguments{
\item{cin}{Cases in CMAM programme}

\item{cout}{Cases not in CMAM programme}

\item{rin}{Recovering cases in CMAM programme}

\item{k}{Correction factor. Ratio of the mean length of an untreated episode
to the mean length of a CMAM treatment episode}
}
\value{
Value of treatment coverage
}
\description{
Estimate treatment coverage
}
\examples{
calculate_tc(cin = 5, cout = 20, rin = 5, k = 3)

}
\references{
Safari Balegamire, Katja Siling, Jose Luis Alvarez Moran,
  Ernest Guevarra, Sophie Woodhead, Alison Norris, Lionella Fieschi,
  Paul Binns, and Mark Myatt (2015). A single coverage estimator for use in
  SQUEAC, SLEAC, and other CMAM coverage assessments. Field Exchange 49,
  March 2015. p81. <www.ennonline.net/fex/49/singlecoverage>
}
\author{
Ernest Guevarra based on technical notes and equations by Mark Myatt
}
