% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monitoring}
\alias{monitoring}
\title{Routine CMAM monitoring data from Sudan}
\format{
A tibble with 8234 rows and 16 columns

| \strong{Variable} | \strong{Description} |
| \emph{State} | Name of state |
| \emph{Locality} | Name of locality |
| \emph{Beginning of Month} | Cases in programme at beginning of month |
| \emph{New Admissions} | New cases admitted within the month |
| \emph{Male} | New male cases admitted within the month |
| \emph{Female} | New female cases admitted within the month |
| \emph{Cured} | Number of cured cases within the month |
| \emph{Death} | Number of cases who died within the month |
| \emph{Default} | Number of cases who defaulted within the month |
| \emph{Non-Responder} | Number of non-responder cases within the month |
| \emph{Total Discharge} | Total number of discharges within the month |
| \emph{RUTF Consumed} | Number of RUTF consumed |
| \emph{Screening} | Screening |
| \emph{Sites} | Sites |
| \emph{Month} | Month |
| \emph{Year} | Year |
}
\source{
Federal Ministry of Health Sudan
}
\usage{
monitoring
}
\description{
Routine CMAM monitoring data from Sudan
}
\keyword{datasets}
