% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{muac_admission}
\alias{muac_admission}
\title{MUAC at admission}
\format{
A tibble with 506 rows and 3 columns

\ \strong{Variable} \ \strong{Description} \
\ \emph{muac} \ Mid-upper arm circumference in centimetres \
\ \emph{district} \ Name of district \
\ \emph{count} \ Number of cases with specific MUAC \
}
\source{
A SQUEAC survey in Lokori, Kenya
}
\usage{
muac_admission
}
\description{
MUAC at admission
}
\examples{
muac_admission

}
\keyword{datasets}
