% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seasonal_calendar}
\alias{seasonal_calendar}
\title{Seasonal calendar data for Sudan}
\format{
A tibble with 28 rows and 5 columns

| \strong{Variables} | \strong{Description} |
| \emph{event} | Name of seasonal calendar event or activity |
| \emph{start} | Starting date of event/activity |
| \emph{end} | Starting date of event/activity |
| \emph{type} | Type of activity or event |
| \emph{group} | Classification/group of activity or event |
}
\source{
https://fews.net/east-africa/sudan/seasonal-calendar/december-2013
}
\usage{
seasonal_calendar
}
\description{
Seasonal calendar data for Sudan
}
\keyword{datasets}
